Folder restructure and cleanup
